Moved StatBlockDivider, renamed folder
