+++ src/lib/dragon/stat-block/bonus-actions/StatBlockBonusActions.svelte
@@ -3,7 +3,7 @@
 	import BActionFrightfulFlare from './BActionFrightfulFlare.svelte';
 	import BActionSupernova from './BActionSupernova.svelte';
 	import BActionVariableRadiance from './BActionVariableRadiance.svelte';
-	import StatBlockDivider from '../StatBlockDivider.svelte';
+	import StatBlockDivider from '$lib/stat-block/StatBlockDivider.svelte';
 	import { dividerHeightThin } from '..';
 
 	export let dragon: DragonStats;
