+++ src/lib/stat-block/StatBlockContainer.svelte
